+++ crates/cli/src/main.rs
@@ -5,6 +5,7 @@ use anyhow::{Context, Result, bail};
 use crossterm::{terminal, tty::IsTty};
 use koto::{
     prelude::*,
+    runtime::{SystemStderr, SystemStdin, SystemStdout},
     serde::{from_koto_value, to_koto_value},
 };
 use koto_format::FormatOptions;
@@ -140,6 +141,10 @@ fn main() -> Result<()> {
         run_tests: args.run_tests || args.run_import_tests,
         vm_settings: KotoVmSettings {
             run_import_tests: args.run_import_tests,
+            args: args.script_args,
+            stdin: make_ptr!(SystemStdin::default()),
+            stdout: make_ptr!(SystemStdout::default()),
+            stderr: make_ptr!(SystemStderr::default()),
             ..Default::default()
         },
     };
@@ -206,7 +211,6 @@ fn main() -> Result<()> {
                             Chunk::instructions_as_string(chunk.clone(), &script_lines)
                         );
                     }
-                    koto.set_args(args.script_args)?;
                     match koto.run(chunk) {
                         Ok(_) => {}
                         Err(error) if error.source().is_some() => {