+++ crates/koto/examples/prelude_value_insert.rs
@@ -6,7 +6,7 @@ print 'name: {name}'
 print 'how_many: {how_many}'
 print 'yes_or_no: {if yes_or_no then 'yes' else 'no'}'
 ";
-    let mut koto = Koto::default();
+    let mut koto = Koto::with_settings(KotoSettings::default().inherit_io());
 
     let prelude = koto.prelude();
     prelude.insert("name", "Alice");