+++ crates/koto/src/koto.rs
@@ -1,6 +1,6 @@
 use crate::{Error, Ptr, Result, prelude::*};
 use koto_bytecode::CompilerSettings;
-use koto_runtime::ModuleImportedCallback;
+use koto_runtime::{ModuleImportedCallback, SystemStderr, SystemStdin, SystemStdout};
 use std::time::Duration;
 
 /// The main interface for the Koto language.
@@ -172,19 +172,6 @@ impl Koto {
         self.runtime.loader().borrow_mut().clear_cache();
     }
 
-    /// Sets the arguments that can be accessed from within the script via `os.args()`
-    pub fn set_args(&mut self, args: impl IntoIterator<Item = String>) -> Result<()> {
-        let koto_args = args.into_iter().map(KValue::from).collect::<Vec<_>>();
-
-        match self.runtime.prelude().data_mut().get("os") {
-            Some(KValue::Map(map)) => {
-                map.insert("args", KValue::Tuple(koto_args.into()));
-                Ok(())
-            }
-            _ => Err(Error::MissingOsModule),
-        }
-    }
-
     /// Enables or disables the `run_tests` setting
     ///
     /// Currently this is only used when running benchmarks where tests are run once during setup,
@@ -203,12 +190,54 @@ pub struct KotoSettings {
 }
 
 impl KotoSettings {
+    /// Helper for conveniently setting the arguments to those of the current process
+    ///
+    /// # Panics
+    ///
+    /// This will panic if any argument of the current process is not valid Unicode.
+    #[must_use]
+    pub fn inherit_args(self) -> Self {
+        Self {
+            vm_settings: KotoVmSettings {
+                args: std::env::args().collect(),
+                ..self.vm_settings
+            },
+            ..self
+        }
+    }
+
+    /// Helper for conveniently setting the stdio streams to those of the current process
+    #[must_use]
+    pub fn inherit_io(self) -> Self {
+        Self {
+            vm_settings: KotoVmSettings {
+                stdin: make_ptr!(SystemStdin::default()),
+                stdout: make_ptr!(SystemStdout::default()),
+                stderr: make_ptr!(SystemStderr::default()),
+                ..self.vm_settings
+            },
+            ..self
+        }
+    }
+
     /// Helper for conveniently defining a maximum execution duration
     #[must_use]
     pub fn with_execution_limit(self, limit: Duration) -> Self {
         Self {
             vm_settings: KotoVmSettings {
                 execution_limit: Some(limit),
+                ..self.vm_settings
+            },
+            ..self
+        }
+    }
+
+    /// Helper for conveniently defining custom args
+    #[must_use]
+    pub fn with_args(self, args: impl IntoIterator<Item: Into<String>>) -> Self {
+        Self {
+            vm_settings: KotoVmSettings {
+                args: args.into_iter().map(Into::into).collect(),
                 ..self.vm_settings
             },
             ..self