+++ crates/runtime/src/core_lib/io.rs
@@ -1,6 +1,9 @@
 //! The `io` core library module
 
-use crate::{BufferedFile, Error, Ptr, Result, derive::*, prelude::*};
+use crate::{
+    BufferedFile, Error, Ptr, Result, UnavailableStderr, UnavailableStdin, UnavailableStdout,
+    derive::*, prelude::*,
+};
 use std::{
     fmt, fs,
     io::{self, BufRead, Read, Seek, SeekFrom, Write},
@@ -126,18 +129,9 @@ pub fn make_module() -> KMap {
         }
     });
 
-    result.add_fn("stderr", |ctx| match ctx.args() {
-        [] => Ok(File::stderr(ctx.vm)),
-        unexpected => unexpected_args("||", unexpected),
-    });
-    result.add_fn("stdin", |ctx| match ctx.args() {
-        [] => Ok(File::stdin(ctx.vm)),
-        unexpected => unexpected_args("||", unexpected),
-    });
-    result.add_fn("stdout", |ctx| match ctx.args() {
-        [] => Ok(File::stdout(ctx.vm)),
-        unexpected => unexpected_args("||", unexpected),
-    });
+    result.insert("stdin", File::new(make_ptr!(UnavailableStdin::default())));
+    result.insert("stdout", File::new(make_ptr!(UnavailableStdout::default())));
+    result.insert("stderr", File::new(make_ptr!(UnavailableStderr::default())));
 
     result.add_fn("temp_dir", |ctx| match ctx.args() {
         [] => Ok(std::env::temp_dir().to_string_lossy().as_ref().into()),
@@ -165,18 +159,6 @@ impl File {
         T: Read + Write + Seek + KotoSend + KotoSync + 'static,
     {
         Self(make_ptr!(BufferedSystemFile::new(file, path))).into()
-    }
-
-    fn stderr(vm: &KotoVm) -> KValue {
-        Self(vm.stderr().clone()).into()
-    }
-
-    fn stdin(vm: &KotoVm) -> KValue {
-        Self(vm.stdin().clone()).into()
-    }
-
-    fn stdout(vm: &KotoVm) -> KValue {
-        Self(vm.stdout().clone()).into()
     }
 
     #[koto_method]