+++ crates/runtime/src/vm.rs
@@ -1,6 +1,7 @@
 use crate::{
-    DefaultStderr, DefaultStdin, DefaultStdout, InstructionFrame, KFunction, Ptr, Result,
-    core_lib::{CoreLib, koto::Unimplemented},
+    InstructionFrame, KFunction, Ptr, Result, UnavailableStderr, UnavailableStdin,
+    UnavailableStdout,
+    core_lib::{CoreLib, io::File, koto::Unimplemented},
     error::{Error, ErrorKind},
     prelude::*,
     types::{FunctionContext, meta_id_to_key, value::RegisterSlice},
@@ -52,6 +53,30 @@ impl VmContext {
     fn with_settings(settings: KotoVmSettings) -> Self {
         let core_lib = CoreLib::default();
 
+        core_lib.os.insert(
+            "args",
+            KValue::Tuple(
+                settings
+                    .args
+                    .iter()
+                    .map(|s| KValue::from(s.as_str()))
+                    .collect::<Vec<_>>()
+                    .into(),
+            ),
+        );
+
+        core_lib
+            .io
+            .insert("stdin", File::new(settings.stdin.clone()));
+
+        core_lib
+            .io
+            .insert("stdout", File::new(settings.stdout.clone()));
+
+        core_lib
+            .io
+            .insert("stderr", File::new(settings.stderr.clone()));
+
         Self {
             settings,
             prelude: core_lib.prelude(),
@@ -95,20 +120,25 @@ pub struct KotoVmSettings {
     /// reload the script when one of its dependencies has changed.
     pub module_imported_callback: Option<Box<dyn ModuleImportedCallback>>,
 
-    /// The runtime's `stdin`
+    /// The runtime's `stdin`that can be accessed from within the script via `io.stdin`
     ///
-    /// Default: [`DefaultStdin`]
+    /// Default: [`UnavailableStdin`]
     pub stdin: Ptr<dyn KotoFile>,
 
-    /// The runtime's `stdout`
+    /// The runtime's `stdout`that can be accessed from within the script via `io.stdout`
     ///
-    /// Default: [`DefaultStdout`]
+    /// Default: [`UnavailableStdout`]
     pub stdout: Ptr<dyn KotoFile>,
 
-    /// The runtime's `stderr`
+    /// The runtime's `stderr` that can be accessed from within the script via `io.stderr`
     ///
-    /// Default: [`DefaultStderr`]
+    /// Default: [`UnavailableStderr`]
     pub stderr: Ptr<dyn KotoFile>,
+
+    /// The runtime's `args` that can be accessed from within the script via `os.args`
+    ///
+    /// Default: `vec![]`
+    pub args: Vec<String>,
 }
 
 impl Default for KotoVmSettings {
@@ -117,9 +147,10 @@ impl Default for KotoVmSettings {
             run_import_tests: true,
             execution_limit: None,
             module_imported_callback: None,
-            stdin: make_ptr!(DefaultStdin::default()),
-            stdout: make_ptr!(DefaultStdout::default()),
-            stderr: make_ptr!(DefaultStderr::default()),
+            stdin: make_ptr!(UnavailableStdin::default()),
+            stdout: make_ptr!(UnavailableStdout::default()),
+            stderr: make_ptr!(UnavailableStderr::default()),
+            args: vec![],
         }
     }
 }