+++ CHANGELOG.md
@@ -76,6 +76,8 @@ The Koto project adheres to
   - Thanks to [@bluurryy](https://github.com/bluurryy) for the contributions.
 - The `KotoAccess` trait has replaced `KotoEntries`, and allows Rust objects to define how `.` access should behave on the object.
 - `KotoVm::run_read_op` and `KotoVm::run_write_op` have been added to run overridden index / access operations.
+- `UnavailableStdin`, `UnavailableStdout` and `UnavailableStderr` have been added to represent unavailable io streams
+- `KotoSettings::inherit_args` and `KotoSettings::inherit_io` have been added to use the args / io of the current process
 
 #### Core Library
 
@@ -84,6 +86,9 @@ The Koto project adheres to
 
 ### Changed
 
+#### Language
+- `io.stdin`, `io.stdout` and `io.stderr` are now provided as `File` instead of `|| -> File`
+
 #### API
 
 - The `KotoEntries` trait has been replaced with `KotoAccess`, see the note in the `Added` section above for more info.
@@ -91,6 +96,9 @@ The Koto project adheres to
   - `BinaryOp::Index` has been moved to `ReadOp::Index`.
   - `MetaKey::IndexMut` has been moved to `WriteOp::IndexAssign`.
   - `KotoObject::index_mut` has been renamed to `KotoObject::index_assign`.
+- `Koto::set_args` has been replaced by `KotoSettings::with_args`
+- `DefaultStdin`, `DefaultStdout` and `DefaultStderr` have been renamed to `SystemStdin`, `SystemStdout` and `SystemStderr`
+- The default io streams of `KotoVmSettings` have been changed to `Unavailable*` io streams
 
 #### Libs